requirements.txt
rembg
pillow
customtkinter
pyinstaller
